creating dtos, reestruturing code and others

+++ src/main/java/com/maua/backendMetro/domain/repository/Localizations.java
@@ -7,5 +7,5 @@ import org.springframework.stereotype.Repository;
 
 @RepositoryRestResource(path = "Localizations")
 @Repository
-public interface Localizations  extends JpaRepository<Localization, Integer> {
+public interface Localizations  extends JpaRepository<Localization, Long> {
 }
