+++ src/main/java/com/maua/backendMetro/rest/controller/dto/ExtinguisherDTO.java
@@ -20,6 +20,6 @@ public class ExtinguisherDTO {
     private ExtinguisherStatus extinguisherStatus;
 
     @NotNull(message = "Localization ID is required")
-    private Integer localizationId;
+    private Long localizationId;
 
 }
